Add files via upload
Tell Me a Joke
This simple JavaScript project lets you request and enjoy a good laugh with just a click of a button. It's a fun and easy way to brighten up your day with humor.

Technologies Used
JavaScript
HTML
CSS

Features
Fetches and displays random jokes.
Simple and intuitive user interface.
A quick way to add a touch of humor to your day.