Add files via upload
This JavaScript project provides an implementation of the Bulb Effect using the Toggle Method. The Bulb Effect is a common visual effect used in various applications to represent the switching on and off of a light bulb. This project demonstrates how to create this effect using the Toggle method in JavaScript.

Technologies Used:

HTML
CSS
Javascript