Add files via upload
This GitHub project contains a comprehensive JavaScript project that implements an Alarm Clock with Timer functionality, with a tune when your alarm rings. The project is designed to provide users with a versatile and customizable solution for setting alarms and timers within web applications. It showcases the effective utilization of JavaScript to create a user-friendly and intuitive alarm and timer interface.

Technologies Used:
HTML
CSS
Javascript